requirements.txt
requests
pandas
zipfile36
bs4
lxml
python-dotenv